Move ChannelAction + PostCollectionConfiguration to `shared` package
