+++ packages/ui/src/components/ChatMessage/ChatMessageActions/MessageActions.tsx
@@ -1,5 +1,6 @@
 import Clipboard from '@react-native-clipboard/clipboard';
 import { type Session, useCurrentSession } from '@tloncorp/shared';
+import { ChannelAction } from '@tloncorp/shared';
 import * as db from '@tloncorp/shared/dist/db';
 import * as logic from '@tloncorp/shared/dist/logic';
 import * as store from '@tloncorp/shared/dist/store';
@@ -81,134 +82,6 @@ function CopyJsonAction({ post }: { post: db.Post }) {
       {!didCopy ? 'Copy post JSON' : 'Copied'}
     </ActionList.Action>
   );
-}
-
-export interface ChannelAction {
-  id: string;
-  label: string;
-  actionType?: 'destructive';
-  networkDependent: boolean;
-}
-export function getPostActions({
-  post,
-  channelType,
-  isMuted,
-}: {
-  post: db.Post;
-  channelType: db.ChannelType;
-  isMuted?: boolean;
-}): ChannelAction[] {
-  switch (channelType) {
-    case 'gallery':
-      return [
-        { id: 'startThread', label: 'Comment on post', networkDependent: true },
-        {
-          id: 'muteThread',
-          label: isMuted ? 'Unmute thread' : 'Mute thread',
-          networkDependent: true,
-        },
-        { id: 'copyRef', label: 'Copy link to post', networkDependent: false },
-        { id: 'edit', label: 'Edit post', networkDependent: true },
-        { id: 'report', label: 'Report post', networkDependent: true },
-        {
-          id: 'visibility',
-          label: post?.hidden ? 'Show post' : 'Hide post',
-          networkDependent: true,
-        },
-        {
-          id: 'delete',
-          label: 'Delete message',
-          actionType: 'destructive',
-          networkDependent: true,
-        },
-      ];
-    case 'notebook':
-      return [
-        { id: 'startThread', label: 'Comment on post', networkDependent: true },
-        {
-          id: 'muteThread',
-          label: isMuted ? 'Unmute thread' : 'Mute thread',
-          networkDependent: true,
-        },
-        { id: 'copyRef', label: 'Copy link to post', networkDependent: false },
-        { id: 'edit', label: 'Edit post', networkDependent: true },
-        { id: 'report', label: 'Report post', networkDependent: true },
-        {
-          id: 'visibility',
-          label: post?.hidden ? 'Show post' : 'Hide post',
-          networkDependent: true,
-        },
-        {
-          id: 'delete',
-          label: 'Delete message',
-          actionType: 'destructive',
-          networkDependent: true,
-        },
-      ];
-    case 'dm':
-    case 'groupDm':
-      return [
-        // { id: 'quote', label: 'Quote' },
-        { id: 'startThread', label: 'Start thread', networkDependent: true },
-        {
-          id: 'muteThread',
-          label: isMuted ? 'Unmute thread' : 'Mute thread',
-          networkDependent: true,
-        },
-        {
-          id: 'viewReactions',
-          label: 'View reactions',
-          networkDependent: false,
-        },
-        { id: 'copyText', label: 'Copy message text', networkDependent: false },
-        {
-          id: 'visibility',
-          label: post?.hidden ? 'Show post' : 'Hide post',
-          networkDependent: true,
-        },
-        {
-          id: 'delete',
-          label: 'Delete message',
-          actionType: 'destructive',
-          networkDependent: true,
-        },
-      ];
-    case 'chat':
-    default:
-      return [
-        { id: 'quote', label: 'Quote', networkDependent: true },
-        { id: 'startThread', label: 'Start thread', networkDependent: true },
-        {
-          id: 'muteThread',
-          label: isMuted ? 'Unmute thread' : 'Mute thread',
-          networkDependent: true,
-        },
-        {
-          id: 'viewReactions',
-          label: 'View reactions',
-          networkDependent: false,
-        },
-        {
-          id: 'copyRef',
-          label: 'Copy link to message',
-          networkDependent: false,
-        },
-        { id: 'copyText', label: 'Copy message text', networkDependent: false },
-        { id: 'edit', label: 'Edit message', networkDependent: true },
-        {
-          id: 'visibility',
-          label: post?.hidden ? 'Show post' : 'Hide post',
-          networkDependent: true,
-        },
-        { id: 'report', label: 'Report message', networkDependent: true },
-        {
-          id: 'delete',
-          label: 'Delete message',
-          actionType: 'destructive',
-          networkDependent: true,
-        },
-      ];
-  }
 }
 
 export async function handleAction({
